fix warning
git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/trunk@1944 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/support/minos/EventServer/gEvServMinos.cxx
@@ -367,7 +367,7 @@ void CalcTotalXSec(string mesg)
   double dE   =   0.200;
   double Emin =   0.010;
   double Emax = 200.010;
-  int    np   = TMath::Ceil((Emax-Emin)/dE);
+  int    np   = (int) TMath::Ceil((Emax-Emin)/dE);
 
   ostringstream xsec_hdr;
   xsec_hdr <<  kXSecCmdSent << ":" << np;
